+++ packages/svelte/src/compiler/phases/3-transform/client/visitors/global.js
@@ -83,8 +83,16 @@ export const global_visitors = {
 			argument.property.type === 'PrivateIdentifier' &&
 			context.state.private_state.has(argument.property.name)
 		) {
-			let fn = node.operator === '++' ? '$.increment' : '$.decrement';
-			return b.call(fn, argument);
+			let fn = '$.update';
+			if (node.prefix) fn += '_pre';
+
+			/** @type {import('estree').Expression[]} */
+			const args = [argument];
+			if (node.operator === '--') {
+				args.push(b.literal(-1));
+			}
+
+			return b.call(fn, ...args);
 		} else {
 			// turn it into an IIFEE assignment expression: i++ -> (() => { const $$value = i; i+=1; return $$value; })
 			const assignment = b.assignment(