+++ packages/svelte/src/compiler/phases/3-transform/server/transform-server.js
@@ -511,24 +511,17 @@ const global_visitors = {
 		const { state, next } = context;
 		const argument = node.argument;
 
-		if (argument.type === 'Identifier') {
-			const binding = state.scope.get(argument.name);
-			const is_store = binding?.kind === 'store_sub';
-			const name = is_store ? argument.name.slice(1) : argument.name;
-			// use runtime functions for smaller output
-			if (is_store) {
-				let fn = node.operator === '++' ? '$.increment' : '$.decrement';
-				if (node.prefix) fn += '_pre';
+		if (argument.type === 'Identifier' && state.scope.get(argument.name)?.kind === 'store_sub') {
+			let fn = '$.update_store';
+			if (node.prefix) fn += '_pre';
 
-				if (is_store) {
-					fn += '_store';
-					return b.call(fn, serialize_get_binding(b.id(name), state), b.call('$' + name));
-				} else {
-					return b.call(fn, b.id(name));
-				}
-			} else {
-				return next();
+			/** @type {import('estree').Expression[]} */
+			const args = [b.id('$$store_subs'), b.literal(argument.name), b.id(argument.name.slice(1))];
+			if (node.operator === '--') {
+				args.push(b.literal(-1));
 			}
+
+			return b.call(fn, ...args);
 		}
 		return next();
 	}
@@ -1452,9 +1445,6 @@ const template_visitors = {
 			context.state.init.push(b.stmt(b.call('$.add_snippet_symbol', node.expression)));
 		}
 	},
-	BindDirective(node, context) {
-		// TODO
-	},
 	Component(node, context) {
 		const state = context.state;
 		const [dec, id] = serialize_anchor(state);
@@ -1694,9 +1684,19 @@ function serialize_element_attributes(node, context) {
 			if (binding?.omit_in_ssr) continue;
 
 			if (ContentEditableBindings.includes(attribute.name)) {
-				content = { escape: false, expression: attribute.expression };
+				content = {
+					escape: false,
+					expression: /** @type {import('estree').Expression} */ (
+						context.visit(attribute.expression)
+					)
+				};
 			} else if (attribute.name === 'value' && node.name === 'textarea') {
-				content = { escape: true, expression: attribute.expression };
+				content = {
+					escape: true,
+					expression: /** @type {import('estree').Expression} */ (
+						context.visit(attribute.expression)
+					)
+				};
 			} else if (attribute.name === 'group') {
 				const value_attribute = /** @type {import('#compiler').Attribute | undefined} */ (
 					node.attributes.find((attr) => attr.type === 'Attribute' && attr.name === 'value')